[style/#255] 사이드바 예전 상태로 되돌리기
